requirements.txt
asgiref==3.4.1
backcall==0.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
decorator==5.1.0
display==1.0.0
dj-database-url==0.5.0
Django==3.2.8
django-bootstrap4==3.0.1
django-heroku==0.3.1
djangorestframework==3.12.4
idna==3.3
ipython==7.28.0
jedi==0.18.0
matplotlib-inline==0.1.3
model-mommy==2.0.0
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.20
psycopg2==2.9.1
Pygments==2.10.0
pytz==2021.3
requests==2.26.0
soupsieve==2.2.1
sqlparse==0.4.2
traitlets==5.1.0
urllib3==1.26.7
wcwidth==0.2.5
whitenoise==5.3.0
